Moved cov to _numpy_compat
