+++ scipy/stats/kde.py
@@ -534,15 +534,9 @@ class gaussian_kde(object):
         self.factor = self.covariance_factor()
         # Cache covariance and inverse covariance of the data
         if not hasattr(self, '_data_inv_cov'):
-            try:
-                self._data_covariance = atleast_2d(np.cov(self.dataset, rowvar=1,
-                                                   bias=False,
-                                                   aweights=self.weights))
-            except TypeError:
-                from scipy.stats._future_cov import cov
-                self._data_covariance = atleast_2d(cov(self.dataset, rowvar=1,
-                                                   bias=False,
-                                                   aweights=self.weights))
+            self._data_covariance = atleast_2d(np.cov(self.dataset, rowvar=1,
+                                               bias=False,
+                                               aweights=self.weights))
             self._data_inv_cov = linalg.inv(self._data_covariance)
 
         self.covariance = self._data_covariance * self.factor**2
